requirements.txt
requests==2.32.3
python-dotenv==1.1.0
imageio==2.37.0
pyrlottie==2024.0.1
rich==14.0.0
